feat: add middleware and guard in module
