+++ core/ctx_test.go
@@ -531,53 +531,6 @@ func Test_Ctx_Status(t *testing.T) {
 	require.Equal(t, http.StatusNotFound, resp.StatusCode)
 }
 
-func Test_CtxContext(t *testing.T) {
-	const key CtxKey = "key"
-
-	middleware := func(ctx Ctx) error {
-		ctx.Set(key, "value")
-		return ctx.Next()
-	}
-	controller := func(module *DynamicModule) *DynamicController {
-		ctrl := module.NewController("test")
-
-		ctrl.Use(middleware).Get("", func(ctx Ctx) {
-			ctx.JSON(Map{
-				"data": ctx.Get(key),
-			})
-		})
-
-		return ctrl
-	}
-
-	module := func() *DynamicModule {
-		appModule := NewModule(NewModuleOptions{
-			Controllers: []Controller{controller},
-		})
-
-		return appModule
-	}
-
-	app := CreateFactory(module)
-	app.SetGlobalPrefix("/api")
-
-	testServer := httptest.NewServer(app.PrepareBeforeListen())
-	defer testServer.Close()
-	testClient := testServer.Client()
-
-	resp, err := testClient.Get(testServer.URL + "/api/test")
-	require.Nil(t, err)
-	require.Equal(t, http.StatusOK, resp.StatusCode)
-
-	data, err := io.ReadAll(resp.Body)
-	require.Nil(t, err)
-
-	var res Response
-	err = json.Unmarshal(data, &res)
-	require.Nil(t, err)
-	require.Equal(t, "value", res.Data)
-}
-
 func Test_QueryParser(t *testing.T) {
 	type QueryData struct {
 		Age    int  `query:"age"`
