+++ core/middleware.go
@@ -20,3 +20,28 @@ func ParseCtxMiddleware(app *App, ctxMid Middleware) middlewareRaw {
 		})
 	}
 }
+
+// Use appends the given middleware functions to the controller's list of
+// middleware handlers. The middleware handlers are run in the order they
+// are added to the controller. The middleware handlers are run before the
+// controller's handlers. The controller's middleware handlers are run
+// after the module's middleware handlers. The module middleware handlers
+// are run after the module's parent middleware handlers. The module
+// middleware handlers are run before the module's controllers. The
+// controller middleware handlers are run before the controller's
+// handlers.
+func (c *DynamicController) Use(middleware ...Middleware) *DynamicController {
+	c.middlewares = append(c.middlewares, middleware...)
+	return c
+}
+
+func (module *DynamicModule) Use(middlewares ...Middleware) *DynamicModule {
+	module.Middlewares = append(module.Middlewares, middlewares...)
+	for _, middleware := range middlewares {
+		for _, router := range module.Routers {
+			router.Middlewares = append(router.Middlewares, middleware)
+		}
+	}
+
+	return module
+}
